requirements.txt
Flask>=2.3.0
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Werkzeug>=2.3.0
